+++ app/app/src/test/java/com/hostshield/util/ParserFuzzHarnessTest.kt
@@ -95,6 +95,7 @@ class ParserFuzzHarnessTest {
             }
             noThrow("dns stamp malformed corpus item $i") {
                 parser.parse(candidate)
+                parser.diagnose(candidate)
             }
         }